+++ src/richie/apps/search/urls.py
@@ -8,6 +8,7 @@ from rest_framework import routers
 from .views import bootstrap_elasticsearch, filter_definitions
 from .viewsets.categories import CategoriesViewSet
 from .viewsets.courses import CoursesViewSet
+from .viewsets.licences import LicencesViewSet
 from .viewsets.organizations import OrganizationsViewSet
 from .viewsets.persons import PersonsViewSet
 
@@ -16,6 +17,7 @@ ROUTER = routers.SimpleRouter()
 
 # Define our app's routes with the router
 ROUTER.register(r"courses", CoursesViewSet, "courses")
+ROUTER.register(r"licences", LicencesViewSet, "licences")
 ROUTER.register(r"organizations", OrganizationsViewSet, "organizations")
 ROUTER.register(r"persons", PersonsViewSet, "persons")
 ROUTER.register(r"(?P<kind>\w+)", CategoriesViewSet, "categories")